# MF0966/examen_mf0966/src/main/java/examen_mf0966/dtos/CursoDTO.java
package examen_mf0966.dtos;


public record CursoDTO (String nombre, String identificador, Integer numHoras, ProfesorDTO profe) {
	public CursoDTO(String nombre, String identificador, Integer numHoras, String nombreProfe, String apellidosProfe) {
		this(nombre, identificador, numHoras, new ProfesorDTO(nombreProfe, apellidosProfe));
	}
}